+++ example/lib/use_closest_build_context_example.dart
@@ -1,4 +1,4 @@
-// ignore_for_file: many_lints/avoid_returning_widgets, many_lints/prefer_overriding_parent_equality, many_lints/prefer_single_widget_per_file
+// ignore_for_file: many_lints/avoid_returning_widgets, many_lints/prefer_overriding_parent_equality, many_lints/prefer_single_widget_per_file, many_lints/never_discard_build_context
 import 'package:flutter/material.dart';
 
 // use_closest_build_context