+++ test/plugin_registration_test.dart
@@ -53,7 +53,7 @@ void main() {
   test('expected number of active and removed rules are registered', () {
     plugin.register(registry);
     final totalRules = registry.warningRules.length + registry.lintRules.length;
-    expect(totalRules, equals(155));
+    expect(totalRules, equals(157));
     expect(
       registry.warningRules.values.whereType<RemovedAnalysisRule>().map(
         (rule) => rule.name,
@@ -165,7 +165,7 @@ void main() {
       0,
       (sum, v) => sum + v.length,
     );
-    expect(totalFixes, equals(90));
+    expect(totalFixes, equals(91));
   });
 
   test('assists are registered', () {